+++ mayan/apps/document_indexing/tests/test_views.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, unicode_literals
 
 from mayan.apps.documents.tests import GenericDocumentViewTestCase
 
-from ..models import Index
+from ..models import Index, IndexInstanceNode
 from ..permissions import (
     permission_document_indexing_create, permission_document_indexing_delete,
     permission_document_indexing_edit,
@@ -10,11 +10,16 @@ from ..permissions import (
     permission_document_indexing_rebuild
 )
 
-from .literals import TEST_INDEX_LABEL, TEST_INDEX_LABEL_EDITED
+from .literals import (
+    TEST_INDEX_LABEL, TEST_INDEX_LABEL_EDITED,
+    TEST_INDEX_TEMPLATE_DOCUMENT_LABEL_EXPRESSION
+)
 from .mixins import IndexTestMixin, IndexViewTestMixin
 
 
 class IndexViewTestCase(IndexTestMixin, IndexViewTestMixin, GenericDocumentViewTestCase):
+    auto_upload_document = False
+
     def test_index_create_view_no_permission(self):
         response = self._request_test_index_create_view()
         self.assertEqual(response.status_code, 403)
@@ -72,6 +77,45 @@ class IndexViewTestCase(IndexTestMixin, IndexViewTestMixin, GenericDocumentViewT
         self.test_index.refresh_from_db()
         self.assertEqual(self.test_index.label, TEST_INDEX_LABEL_EDITED)
 
+    def test_index_rebuild_view_no_permission(self):
+        self.upload_document()
+        self._create_test_index()
+
+        self.test_index.node_templates.create(
+            parent=self.test_index.template_root,
+            expression=TEST_INDEX_TEMPLATE_DOCUMENT_LABEL_EXPRESSION,
+            link_documents=True
+        )
+
+        response = self._request_test_index_rebuild_view()
+        self.assertEqual(response.status_code, 404)
+
+        self.assertEqual(IndexInstanceNode.objects.count(), 0)
+
+    def test_index_rebuild_view_with_access(self):
+        self.upload_document()
+        self._create_test_index()
+
+        self.test_index.node_templates.create(
+            parent=self.test_index.template_root,
+            expression=TEST_INDEX_TEMPLATE_DOCUMENT_LABEL_EXPRESSION,
+            link_documents=True
+        )
+
+        self.grant_access(
+            obj=self.test_index,
+            permission=permission_document_indexing_rebuild
+        )
+
+        response = self._request_test_index_rebuild_view()
+        self.assertEqual(response.status_code, 302)
+
+        self.assertNotEqual(IndexInstanceNode.objects.count(), 0)
+
+
+class IndexInstaceViewTestCase(
+    IndexTestMixin, IndexViewTestMixin, GenericDocumentViewTestCase
+):
     def _request_index_instance_node_view(self, index_instance_node):
         return self.get(
             viewname='indexing:index_instance_node_view', kwargs={
@@ -100,9 +144,13 @@ class IndexViewTestCase(IndexTestMixin, IndexViewTestMixin, GenericDocumentViewT
         )
         self.assertContains(response, text=TEST_INDEX_LABEL, status_code=200)
 
+
+class IndexToolsViewTestCase(
+    IndexTestMixin, IndexViewTestMixin, GenericDocumentViewTestCase
+):
     def _request_index_rebuild_get_view(self):
         return self.get(
-            viewname='indexing:rebuild_index_instances',
+            viewname='indexing:rebuild_index_instances'
         )
 
     def _request_index_rebuild_post_view(self):
@@ -112,7 +160,7 @@ class IndexViewTestCase(IndexTestMixin, IndexViewTestMixin, GenericDocumentViewT
             }
         )
 
-    def test_index_rebuild_no_permission(self):
+    def test_indexes_rebuild_no_permission(self):
         self._create_test_index(rebuild=False)
 
         response = self._request_index_rebuild_get_view()
@@ -128,7 +176,7 @@ class IndexViewTestCase(IndexTestMixin, IndexViewTestMixin, GenericDocumentViewT
             self.test_index.instance_root.get_children_count(), 0
         )
 
-    def test_index_rebuild_with_access(self):
+    def test_indexes_rebuild_with_access(self):
         self._create_test_index(rebuild=False)
 
         self.grant_access(